refactor(FavoriteLocation): Replace isFixed prop with placeholder prop
