+++ lib/components/user/favorite-location.js
@@ -9,7 +9,6 @@ import {
 import styled from 'styled-components'
 
 import Icon from '../narrative/icon'
-import { isHome, isWork } from '../../util/user'
 import { FavoriteLocationField, FavoriteLocationTypeDropdown } from './favorite-location-controls'
 
 // Styles for other controls
@@ -21,24 +20,10 @@ const StyledAddon = styled(InputGroup.Addon)`
   width: 40px;
 `
 
-// Defaults for home and work
-export const BLANK_HOME = {
-  address: '',
-  icon: 'home',
-  name: '',
-  type: 'home'
-}
-export const BLANK_WORK = {
-  address: '',
-  icon: 'briefcase',
-  name: '',
-  type: 'work'
-}
-
 /**
  * Creates a new favorite location object from a location returned by LocationField, with the specified type and icon.
  */
-function makeFavoriteLocation (baseLocation, type, icon) {
+export function makeFavoriteLocation (baseLocation, type, icon) {
   const { lat, lon, name } = baseLocation
   return {
     address: name,
@@ -55,25 +40,10 @@ function makeFavoriteLocation (baseLocation, type, icon) {
  * and lets the user edit the details (address, type, TODO: nickname) of the location.
  */
 class FavoriteLocation extends Component {
-  constructor () {
-    super()
-    this.state = {
-      newLocation: null
-    }
-  }
-
   _handleDelete = () => {
-    const { arrayHelpers, index, isFixed, location } = this.props
-    if (isFixed) {
-      // For 'Home' and 'Work', replace with the default blank locations instead of deleting.
-      let newLocation
-      if (isHome(location)) {
-        newLocation = BLANK_HOME
-      } else if (isWork(location)) {
-        newLocation = BLANK_WORK
-      }
-
-      this._handleLocationChange({ location: newLocation })
+    const { arrayHelpers, index, placeholder } = this.props
+    if (placeholder) {
+      this._handleLocationChange({ location: placeholder })
     } else {
       arrayHelpers.remove(index)
     }
@@ -90,25 +60,14 @@ class FavoriteLocation extends Component {
     arrayHelpers.replace(index, makeFavoriteLocation(location, type, icon))
   }
 
-  _handleNew = ({ location }) => {
-    const { arrayHelpers } = this.props
-
-    // Set, then unset the location state to reset the LocationField after address is entered.
-    // (both this.setState calls should trigger componentDidUpdate on LocationField).
-    this.setState({ newLocation: location })
-
-    arrayHelpers.push(makeFavoriteLocation(location, 'custom', 'map-marker'))
-
-    this.setState({ newLocation: null })
-  }
-
   render () {
-    const { index, isFixed, location } = this.props
-    const placeholder = isFixed ? `Add ${location.type}` : 'Enter a favorite address'
+    const { index, location, placeholder } = this.props
+    const placeholderText = placeholder ? `Add ${location.type}` : 'Enter a favorite address'
 
-    // Show a dropdown for editing the icon, unless isFixed = true (for 'home' and 'work' locations).
+    // Show a dropdown for editing the icon, shown unless a placeholder is set,
+    // in which case the type is considered fixed (e.g. for 'home' and 'work' locations).
     // The dropdown has a predefined list of items for location types.
-    const iconControl = isFixed
+    const iconControl = placeholder
       ? (
         <StyledAddon title={location.type}>
           <Icon fixedWidth={false} name={location.icon} />
@@ -130,7 +89,7 @@ class FavoriteLocation extends Component {
           {/* wrapper with z-index override needed for showing location menu on top of other controls. */}
           <span className='form-control' style={{ padding: 0, zIndex: 'inherit' }}>
             <FavoriteLocationField
-              inputPlaceholder={placeholder}
+              inputPlaceholder={placeholderText}
               location={location}
               locationType='to'
               onLocationSelected={this._handleLocationChange}
