+++ lib/components/user/favorite-locations-pane.js
@@ -4,15 +4,31 @@ import { ControlLabel } from 'react-bootstrap'
 
 import FavoriteLocation from './favorite-location'
 import NewFavoriteLocation from './new-favorite-location'
-import { isHome, isWork } from '../../util/user'
+
+// Defaults for home and work
+export const BLANK_HOME = {
+  address: '',
+  icon: 'home',
+  name: '',
+  type: 'home'
+}
+export const BLANK_WORK = {
+  address: '',
+  icon: 'briefcase',
+  name: '',
+  type: 'work'
+}
+
+const placeholders = {
+  home: BLANK_HOME,
+  work: BLANK_WORK
+}
 
 /**
  * Renders an editable list user's favorite locations, and lets the user add a new one.
  */
 const FavoriteLocationsPane = ({ values: userData }) => {
   const { savedLocations } = userData
-  const homeLocation = savedLocations.find(isHome)
-  const workLocation = savedLocations.find(isWork)
 
   return (
     <div>
@@ -22,18 +38,15 @@ const FavoriteLocationsPane = ({ values: userData }) => {
         name='savedLocations'
         render={arrayHelpers => (
           <>
-            {savedLocations.map((loc, index) => {
-              const isHomeOrWork = loc === homeLocation || loc === workLocation
-              return (
-                <FavoriteLocation
-                  arrayHelpers={arrayHelpers}
-                  index={index}
-                  isFixed={isHomeOrWork}
-                  key={index}
-                  location={loc}
-                />
-              )
-            })}
+            {savedLocations.map((loc, index) => (
+              <FavoriteLocation
+                arrayHelpers={arrayHelpers}
+                index={index}
+                key={index}
+                location={loc}
+                placeholder={placeholders[loc.type]}
+              />
+            ))}
 
             {/* For adding a location. */}
             <NewFavoriteLocation arrayHelpers={arrayHelpers} />
