+++ lib/components/user/new-favorite-location.js
@@ -6,6 +6,7 @@ import {
 } from 'react-bootstrap'
 import styled from 'styled-components'
 
+import { makeFavoriteLocation } from './favorite-location'
 import { FavoriteLocationField, FavoriteLocationTypeDropdown } from './favorite-location-controls'
 
 // Styles
@@ -14,21 +15,6 @@ const NewLocationDropdownButton = styled(DropdownButton)`
   color: #fff;
   width: 40px;
 `
-
-/**
- * Creates a new favorite location object from a location returned by LocationField, with the specified type and icon.
- */
-function makeFavoriteLocation (baseLocation, type, icon) {
-  const { lat, lon, name } = baseLocation
-  return {
-    address: name,
-    icon,
-    lat,
-    lon,
-    name,
-    type
-  }
-}
 
 const blankNewLocationState = {
   newLocation: null,
