+++ lib/components/user/user-account-screen.js
@@ -12,8 +12,7 @@ import { isNewUser, isHome, isWork } from '../../util/user'
 import DesktopNav from '../app/desktop-nav'
 import AccountSetupFinishPane from './account-setup-finish-pane'
 import ExistingAccountDisplay from './existing-account-display'
-import { BLANK_HOME, BLANK_WORK } from './favorite-location'
-import FavoriteLocationsPane from './favorite-locations-pane'
+import FavoriteLocationsPane, { BLANK_HOME, BLANK_WORK } from './favorite-locations-pane'
 import NewAccountWizard from './new-account-wizard'
 import NotificationPrefsPane from './notification-prefs-pane'
 import TermsOfUsePane from './terms-of-use-pane'
